フォームをmoleculesに移行し,Loginページ,SignUpページをコンポーネントを使って実装しました.
