+++ libs/src/components/molecules/Form/story.tsx
@@ -1,0 +1,13 @@
+import Form from "."
+import React from "react"
+
+export default {
+    title: "Molecules/Form",
+    component: Form,
+}
+
+export const Forms = () => {
+    return(
+        <Form text={"Login"} action={"login"}></Form>
+    )
+}
